New descriptor (density) and new scoring function + other minor modifications.
Also, descall.sh is a script that  launch dpocket on all dataset available, and store results in a directory named using parameters defined in the script. A third file is created, which merge both pocket and non-pocket output files in a single one named all_dataset.txt.

The same idea is used by testall.sh, but instead we launch tpocket.